xuiApi/server/backend/requirements.txt
Django==4.1.6
djangorestframework==3.14.0
django-sequences==2.7
requests


